+++ source/libmodem/proto/at/at_common.h
@@ -56,15 +56,15 @@ modem_fw_ver_t* at_get_fw_version(modem_t* modem, modem_fw_ver_t* fw_info);
 
 /*------------------------------------------------------------------------*/
 
-char* mc77x0_at_get_network_type(modem_t* modem, char* network, int len);
+char* mc77x0_at_get_network_type(modem_t* modem, char* network, size_t len);
 
 /*------------------------------------------------------------------------*/
 
-char* at_get_operator_name(modem_t* modem, char* oper, int len);
+char* at_get_operator_name(modem_t* modem, char* oper, size_t len);
 
 /*------------------------------------------------------------------------*/
 
-char* at_get_operator_number(modem_t* modem, char* oper_number, int len);
+char* at_get_operator_number(modem_t* modem, char* oper_number, size_t len);
 
 /*------------------------------------------------------------------------*/
